ArgParser revised
Add a -r (--root) parameter. This searches for and parses all yaml specs in that path, recursively.
If the -s (--spec) parameter is provided, the explicit file will instead be used. When --root is used,
the results are searched for the first spec file that can handle the extension of the binary.

Cleaned up the test/example files